requirements.txt
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
python-dotenv==0.19.1
